pyproject.toml
[project]
name = "mini-data-platform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-identity>=1.23.0",
    "azure-storage-file-datalake>=12.20.0",
    "deltalake>=0.25.5",
    "loguru>=0.7.3",
    "pandas>=2.2.3",
    "pyarrow>=20.0.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
